Decouple std::hash specializations for std types
